# fix(atoms): fix and optimize quicksort calculations