--- a/packages/atoms/src/classes/schedulers/AsyncScheduler.ts
+++ b/packages/atoms/src/classes/schedulers/AsyncScheduler.ts
@@ -1,27 +1,13 @@
 import { Job } from '@zedux/atoms/types/index'
 import { runJobs, SchedulerBase } from './SchedulerBase'
 
+// it's fine if we "leak" microtasks by queueing one then flushing
+// synchronously. The microtask will just be a noop. And it's a microtask, so
+// not a real leak unlike with `setTimeout`
 const queueMicrotask =
   globalThis.queueMicrotask || (cb => Promise.resolve().then(cb))
 
-const schedule = (scheduler: AsyncScheduler) => {
-  if (scheduler.r) return
-
-  queueMicrotask(() => {
-    runJobs(scheduler)
-  })
-}
-
 export class AsyncScheduler extends SchedulerBase {
-  /**
-   * Kill any current timeout and run all jobs immediately.
-   */
-  public flush() {
-    if (this.r) return // already flushing
-
-    runJobs(this)
-  }
-
   /**
    * Insert a job at the end of the queue. Schedule a flush if this scheduler is
    * not currently running and not scheduled yet.
@@ -31,9 +17,7 @@ export class AsyncScheduler extends SchedulerBase {
    * render). We use the `scheduleAsync` helper instead.
    */
   public schedule(newJob: Job) {
-    // schedule if we just pushed the first job onto the queue and aren't
-    // already running (can happen if an effect immediately triggers the sync
-    // scheduler which schedules more async jobs).
-    this.j.push(newJob) === 1 && !this.r && schedule(this)
+    // schedule if we just pushed the first job onto the queue
+    this.j.push(newJob) === 1 && queueMicrotask(() => runJobs(this))
   }
 }