## Description

#203 introduced a bug in the quicksort algorithm (totally my bad - the equivalent of `2 ** iteration` before is `divisor * 2`, not `divisor * divisor` 🤦 ). This fortunately made me realize we don't have tests for that algorithm. So fix the problem and add tests.

While I was in there, I played with tweaking the maths slightly and actually procured a 10-15% speed boost. The `broadPropagation` benchmark (the one that utilizes the scheduler the most) is down from ~760 ms to ~690 ms. `avoidablePropagation` and `deepPropagation` are almost 20% faster.

All of this has led me to realize that the O(n) push-pull algorithm being O(2n) is actually a big deal here. We'd typically say 2n and n are the same big O, but when you're comparing to O(n log n), they're not. 2n is always worse than n log n. The only reason Zedux's scheduling isn't faster than push-pull-based signals libs is because we can't control array resizing. Our splice call alone is 80% of the total job scheduling time. And job scheduling is 25% of total execution time in the benchmarks.

Regardless, despite being 10% slower at broad propagation, Zedux is now about 10% faster than SolidJS at running the full benchmark suite (27.6s vs 31s). I am still planning on switching to a push-pull model.